chore: rename toggle-group-element to toggle-group-button
