+++ libs/components/src/toggle-group/toggle-group.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from "@angular/core";
 
 import { BadgeModule } from "../badge";
 
-import { ToggleGroupElementComponent } from "./toggle-group-element.component";
+import { ToggleGroupElementComponent } from "./toggle-group-button.component";
 import { ToggleGroupComponent } from "./toggle-group.component";
 
 @NgModule({
